+++ src/pages/Dashboard/Dashboard.tsx
@@ -1,8 +1,18 @@
-import "./Dashboard.scss";
-import { Card, Col, Row, Image, Container, ProgressBar } from "react-bootstrap";
-import WorldMapComponent from "./WorldMapComponent";
+import React from "react";
+import { Card, Col, Container, Row } from "react-bootstrap";
+import WelcomeCard from "./WelcomeCard/WelcomeCard";
+import DashboardStatCard from "./DashboardStatCard/DashboardStatCard";
+import SocialLoginBar from "./SocialLoginBar/SocialLoginBar";
+import DonutChart from "./DonutChart/DonutChart";
+import ActivityLineChart from "./ActivityLineChart/ActivityLineChart";
+import RecentUsers from "./RecentUsers/RecentUsers";
+import WorldMapComponent from "./WorldMapComponent/WorldMapComponent";
 
-//---------------------------------------------
+const stats = {
+  UserName: "Mathew Anderson",
+  TotalUsers: 1200,
+  ActiveUsers: 850,
+};
 
 const stats1 = [
   {
@@ -31,45 +41,14 @@ const stats1 = [
   },
 ];
 
-// ------------------------------------------
-
-const stats = {
-  UserName: "Mathew Anderson",
-  TotalUsers: 1200,
-  ActiveUsers: 850,
-};
-
-// ------------------------------------------
 const socialLogins = [
-  { type: "Google", value: 90 },
-  { type: "Facebook", value: 30 },
-  { type: "Github", value: 20 },
-  { type: "Twitter", value: 10 },
-  { type: "Linkedin", value: 5 },
+  { type: "Google", value: 90, color: "custom-0" },
+  { type: "Facebook", value: 30, color: "custom-1" },
+  { type: "Github", value: 20, color: "custom-2" },
+  { type: "Twitter", value: 10, color: "custom-3" },
+  { type: "Linkedin", value: 5, color: "custom-4" },
+  { type: "Instagram", value: 5, color: "custom-5" },
 ];
-
-const SocialLoginBar = ({
-  type,
-  value,
-  color,
-}: {
-  type: string;
-  value: number;
-  color: string;
-}) => (
-  <div className="social-login-bar mb-3">
-    <div className="d-flex justify-content-between align-items-center mb-1">
-      <span className="login-type">
-        <i className={`bi bi-${type.toLowerCase()}`}></i>
-        {type}
-      </span>
-      <span className="login-value">{value}%</span>
-    </div>
-    <ProgressBar now={value} variant={color} className="custom-progress" />
-  </div>
-);
-
-// ------------------------------------------
 
 const browserStats = [
   { name: "Chrome", value: 70, color: "#4318FF" },
@@ -78,223 +57,43 @@ const browserStats = [
   { name: "Edge", value: 10, color: "#1877F2" },
 ];
 
-const DonutChart = ({ data }: { data: typeof browserStats }) => {
-  const total = data.reduce((sum, item) => sum + item.value, 0);
-  let rotateOffset = 0;
-
-  return (
-    <div className="donut-chart-container">
-      <div className="donut-chart">
-        {data.map((item, index) => {
-          const degrees = (item.value / total) * 360;
-          const style = {
-            "--offset": `${rotateOffset}deg`,
-            "--value": `${degrees}deg`,
-            "--bg-color": item.color,
-          } as React.CSSProperties;
-          rotateOffset += degrees;
-          return (
-            <div key={item.name} className="donut-segment" style={style} />
-          );
-        })}
-      </div>
-      <div className="donut-legend">
-        {data.map((item) => (
-          <div key={item.name} className="legend-item">
-            <span
-              className="legend-dot"
-              style={{ backgroundColor: item.color }}
-            />
-            <span className="legend-label">{item.name}</span>
-            <span className="legend-value">{item.value}%</span>
-          </div>
-        ))}
-      </div>
-    </div>
-  );
-};
-//------------------------------------------
-
-const ActivityLineChart = () => {
-  const hours = ["12am", "3am", "6am", "9am", "12pm", "3pm", "6pm", "9pm"];
-  const values = [30, 20, 25, 85, 90, 70, 95, 60];
-
-  const maxValue = Math.max(...values);
-  const points = values
-    .map((value, index) => {
-      const x = (index / (values.length - 1)) * 100;
-      const y = 100 - (value / maxValue) * 100;
-      return `${x},${y}`;
-    })
-    .join(" ");
-
-  return (
-    <div className="activity-chart">
-      <div className="chart-labels y-labels">
-        {[maxValue, maxValue * 0.75, maxValue * 0.5, maxValue * 0.25, 0].map(
-          (label, index) => (
-            <span key={index}>{Math.round(label)}</span>
-          )
-        )}
-      </div>
-      <div className="chart-container">
-        <svg viewBox="0 0 100 100" preserveAspectRatio="none">
-          <defs>
-            <linearGradient id="lineGradient" x1="0%" y1="0%" x2="100%" y2="0%">
-              <stop offset="0%" stopColor="#4318FF" />
-              <stop offset="100%" stopColor="#9F7AEA" />
-            </linearGradient>
-            <linearGradient id="areaGradient" x1="0%" y1="0%" x2="0%" y2="100%">
-              <stop offset="0%" stopColor="#4318FF" stopOpacity="0.3" />
-              <stop offset="100%" stopColor="#4318FF" stopOpacity="0.02" />
-            </linearGradient>
-          </defs>
-
-          <path
-            d={`M0,100 L${points} L100,100 Z`}
-            fill="url(#areaGradient)"
-            className="area-path"
-          />
-
-          <polyline
-            points={points}
-            fill="none"
-            stroke="url(#lineGradient)"
-            strokeWidth="0."
-            className="main-line"
-          />
-        </svg>
-        <div className="x-labels">
-          {hours.map((hour, index) => (
-            <span key={index}>{hour}</span>
-          ))}
-        </div>
-      </div>
-    </div>
-  );
-};
-
-//----------------------------------------
-
 const recentUsers = [
-  {
-    date: "11.2.2021",
-    name: "John Smith",
-  },
-  {
-    date: "11.2.2021",
-    name: "Sarah Johnson",
-  },
-  {
-    date: "11.2.2021",
-    name: "Michael Brown",
-  },
-  {
-    date: "11.2.2021",
-    name: "Emma Wilson",
-  },
-  {
-    date: "11.2.2021",
-    name: "James Davis",
-  },
-  {
-    date: "11.2.2021",
-    name: "James Davis",
-  },
+  { date: "11.2.2021", name: "John Smith" },
+  { date: "11.2.2021", name: "Sarah Johnson" },
+  { date: "11.2.2021", name: "Michael Brown" },
+  { date: "11.2.2021", name: "Emma Wilson" },
+  { date: "11.2.2021", name: "James Davis" },
+  { date: "11.2.2021", name: "Olivia Martinez" },
+];
+const activityData = {
+  hours: ["12am", "3am", "6am", "9am", "12pm", "3pm", "6pm", "9pm"],
+  values: [30, 20, 25, 85, 90, 70, 95, 60],
+};
+
+const users = [
+  { coordinates: [10, 50] as [number, number] },
+  { coordinates: [-122.4194, 37.7749] as [number, number] },
+  { coordinates: [-74.006, 40.7128] as [number, number] },
+  { coordinates: [139.6917, 35.6895] as [number, number] },
+  { coordinates: [77.209, 28.6139] as [number, number] },
 ];
 
-const RecentUsers = () => {
-  return (
-    <div className="recent-users-list px-5">
-      {recentUsers.map((user, index) => (
-        <div key={index} className="user-timeline-item">
-          <div className="time-stamp">{user.date}</div>
-          <div className="user-info">
-            <div className="timeline-dot"></div>
-            <div className="user-name">{user.name}</div>
-          </div>
-        </div>
-      ))}
-    </div>
-  );
-};
 const Dashboard = () => {
   return (
     <Container fluid className="pt-4">
-      <Row className="mb-4 ">
-        <Col xs={12}>
-          <Card
-            className="welcome-card rounded-4 border-0 shadow-sm"
-            style={{ backgroundColor: "#eef3ff" }}
-          >
-            <Card.Body className="p-0 px-2">
-              <Row className="align-items-center">
-                <Col xs={12} md={7} className="welcome-content  p-md-5 ">
-                  <div className="welcome-header mb">
-                    <h5 className="fw-semibold text-dark mb-3 fs-2">
-                      Welcome back, {stats.UserName}!
-                    </h5>
-                  </div>
-                  <div className="stats-container d-flex flex-column flex-md-row">
-                    <div className="stat-item text-center mb-3 mb-md-0  ">
-                      <h3 className="mt-3 fw-bold text-dark fs-5 d-flex align-items-center justify-content-center">
-                        {stats.TotalUsers}
-                      </h3>
-                      <p className="mb-0 text-secondary fw-medium ">
-                        Total users
-                      </p>
-                    </div>
-                    <div className="stat-item text-center ps-md-5">
-                      <h3 className="mt-3 fw-bold text-dark fs-5 d-flex align-items-center justify-content-center">
-                        {stats.ActiveUsers}
-                      </h3>
-                      <p className="mb-0 text-secondary fw-medium">
-                        Active users
-                      </p>
-                    </div>
-                  </div>
-                </Col>
-                <Col
-                  xs={12}
-                  md={5}
-                  className="welcome-illustration text-center"
-                >
-                  <Image
-                    src="https://bootstrapdemos.adminmart.com/modernize/dist/assets/images/backgrounds/welcome-bg.svg"
-                    alt="Welcome Illustration"
-                    className="img-fluid"
-                  />
-                </Col>
-              </Row>
-            </Card.Body>
-          </Card>
-        </Col>
+      <Row className="mb-4">
+        <WelcomeCard stats={stats} />
       </Row>
 
       <Row className="mb-4">
-        {Object.entries(stats1).map(([key, value], index) => (
-          <Col xs={12} sm={6} md={3} lg={3} key={index} className="mb-3">
-            <Card
-              className={`dashboard-stat-card border-0 shadow-sm bg-light-${value.color}`}
-            >
-              <Card.Body>
-                <div className="text-center">
-                  <div className={`icon-circle mb-3 `}>
-                    <Image
-                      src={value.img}
-                      className={`bi bi-person fs-4 text-${value.color}`}
-                    />
-                  </div>
-                  <p className={`fw-semibold fs-5 text-${value.color} mb-1`}>
-                    {value.title}
-                  </p>
-                  <h5 className={`fw-semibold fs-5 text-${value.color} mb-0`}>
-                    {value.value}
-                  </h5>
-                </div>
-              </Card.Body>
-            </Card>
-          </Col>
+        {stats1.map((stat, index) => (
+          <DashboardStatCard
+            key={index}
+            title={stat.title}
+            value={stat.value}
+            img={stat.img}
+            color={stat.color}
+          />
         ))}
       </Row>
 
@@ -306,7 +105,7 @@ const Dashboard = () => {
           >
             <Card.Body className="px-5">
               <div className="d-flex align-items-center mb-2">
-                <h5 className="card-title mt-2 ">Social Login Statistics</h5>
+                <h5 className="card-title mt-2">Social Login Statistics</h5>
               </div>
 
               <div className="social-login-bars">
@@ -315,7 +114,7 @@ const Dashboard = () => {
                     key={login.type}
                     type={login.type}
                     value={login.value}
-                    color={`custom-${index}`}
+                    color={login.color}
                   />
                 ))}
               </div>
@@ -324,7 +123,7 @@ const Dashboard = () => {
         </Col>
         <Col xs={12} md={4}>
           <Card
-            className="browser-stats-card border-0 shadow-sm  "
+            className="browser-stats-card border-0 shadow-sm"
             style={{ backgroundColor: "#eef3ff" }}
           >
             <Card.Body>
@@ -346,7 +145,7 @@ const Dashboard = () => {
               <div className="d-flex align-items-center">
                 <h5 className="card-title mb-0">Recent Users</h5>
               </div>
-              <RecentUsers />
+              <RecentUsers users={recentUsers} />
             </Card.Body>
           </Card>
         </Col>
@@ -359,7 +158,10 @@ const Dashboard = () => {
               <div className="d-flex justify-content-between align-items-center mb-4">
                 <h5 className="card-title mb-0">User Activity Hours</h5>
               </div>
-              <ActivityLineChart />
+              <ActivityLineChart
+                hours={activityData.hours}
+                values={activityData.values}
+              />
             </Card.Body>
           </Card>
         </Col>
@@ -367,7 +169,7 @@ const Dashboard = () => {
 
       <Row className="mb-4">
         <Col xs={12} md={12} lg={12}>
-          <WorldMapComponent />
+          <WorldMapComponent users={users} />
         </Col>
       </Row>
     </Container>
