+++ src/pages/Dashboard/DonutChart/DonutChart.scss
@@ -1,0 +1,60 @@
+.donut-chart-container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  padding: 2rem 3rem;
+  .donut-chart {
+    position: relative;
+    width: 120px;
+    height: 120px;
+    margin-bottom: 1rem;
+    border-radius: 50%;
+    background: #fff;
+    .donut-segment {
+      position: absolute;
+      width: 100%;
+      height: 100%;
+      border-radius: 50%;
+      transform: rotate(var(--offset));
+      background: conic-gradient(
+        var(--bg-color) var(--value),
+        transparent var(--value)
+      );
+    }
+    &::after {
+      content: "";
+      position: absolute;
+      width: 80%;
+      height: 80%;
+      background: #eef3ff;
+      border-radius: 50%;
+      top: 10%;
+      left: 10%;
+    }
+  }
+  .donut-legend {
+    width: 100%;
+    .legend-item {
+      display: flex;
+      align-items: center;
+      margin-bottom: 0.25rem;
+      padding: 0.15rem 0;
+      .legend-dot {
+        width: 8px;
+        height: 8px;
+        border-radius: 50%;
+        margin-right: 8px;
+      }
+      .legend-label {
+        flex: 1;
+        color: #64748b;
+        font-size: 0.875rem;
+      }
+      .legend-value {
+        color: #1e293b;
+        font-weight: 600;
+        font-size: 0.875rem;
+      }
+    }
+  }
+}
