+++ src/pages/Dashboard/RecentUsers/RecentUsers.tsx
@@ -1,0 +1,24 @@
+import React from "react";
+import "./RecentUsers.scss";
+
+interface RecentUsersProps {
+  users: Array<{ date: string; name: string }>;
+}
+
+const RecentUsers = ({ users }: RecentUsersProps) => {
+  return (
+    <div className="recent-users-list px-5">
+      {users.map((user, index) => (
+        <div key={index} className="user-timeline-item">
+          <div className="time-stamp">{user.date}</div>
+          <div className="user-info">
+            <div className="timeline-dot"></div>
+            <div className="user-name">{user.name}</div>
+          </div>
+        </div>
+      ))}
+    </div>
+  );
+};
+
+export default RecentUsers;
