+++ src/pages/Dashboard/WorldMapComponent/WorldMapComponent.scss
@@ -1,0 +1,29 @@
+.world-map-card {
+  .card-title {
+    color: #2c3e50;
+    font-size: 1.1rem;
+    font-weight: 600;
+  }
+
+  .map-container {
+    width: 100%;
+    position: relative;
+    overflow: hidden;
+  }
+
+  .rsm-geography {
+    fill: #002851;
+    stroke: #65a8e6;
+    stroke-width: 0.5;
+    outline: none;
+
+    &:hover {
+      fill: #97c5ef;
+      cursor: pointer;
+    }
+  }
+
+  .rsm-marker circle {
+    fill: #f07057;
+  }
+}
